Move storeCurrentState to where it belongs
