+++ mapping/mappings.go
@@ -1,11 +1,9 @@
 package mapping
 
 import (
-	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"os/user"
-	"path"
 	"path/filepath"
 )
 
@@ -117,26 +115,4 @@ func loadAllMappings() ([]Mapping, error) {
 
 	sortMappings(mappings)
 	return mappings, nil
-}
-
-func storeCurrentState() error {
-	toStore := make(map[string]Mapping)
-	for _, mapping := range mappings {
-		toStore[mapping.MappingID] = mapping
-	}
-
-	// Force reload
-	mappings = nil
-
-	data, dataErr := json.Marshal(toStore)
-	if dataErr != nil {
-		return dataErr
-	}
-
-	mappingDir, mappingDirErr := getMappingDirectory()
-	if mappingDirErr != nil {
-		return mappingDirErr
-	}
-
-	return ioutil.WriteFile(path.Join(mappingDir, stateFileName), data, 0644)
 }
